+++ packages/backend/src/App.ts
@@ -1,0 +1,414 @@
+import { LightdashMode, SessionUser } from '@lightdash/common';
+import { NodeSDK } from '@opentelemetry/sdk-node';
+import * as Sentry from '@sentry/node';
+import * as Tracing from '@sentry/tracing';
+import { SamplingContext } from '@sentry/types';
+import bodyParser from 'body-parser';
+import flash from 'connect-flash';
+import connectSessionKnex from 'connect-session-knex';
+import cookieParser from 'cookie-parser';
+import express, {
+    Express,
+    NextFunction,
+    Request,
+    RequestHandler,
+    Response,
+} from 'express';
+import expressSession from 'express-session';
+import expressStaticGzip from 'express-static-gzip';
+import passport from 'passport';
+import refresh from 'passport-oauth2-refresh';
+import path from 'path';
+import reDoc from 'redoc-express';
+import { LightdashAnalytics } from './analytics/LightdashAnalytics';
+import * as clients from './clients/clients';
+import { SlackService } from './clients/Slack/Slackbot';
+import { LightdashConfig } from './config/parseConfig';
+import {
+    apiKeyPassportStrategy,
+    azureAdPassportStrategy,
+    googlePassportStrategy,
+    isOktaPassportStrategyAvailableToUse,
+    localPassportStrategy,
+    oneLoginPassportStrategy,
+    OpenIDClientOktaStrategy,
+} from './controllers/authentication';
+import database from './database/database';
+import { errorHandler } from './errors';
+import { RegisterRoutes } from './generated/routes';
+import apiSpec from './generated/swagger.json';
+import Logger from './logging/logger';
+import { expressWinstonMiddleware } from './logging/winston';
+import { slackAuthenticationModel, userModel } from './models/models';
+import { registerNodeMetrics } from './nodeMetrics';
+import { postHogClient } from './postHog';
+import { apiV1Router } from './routers/apiV1Router';
+import { SchedulerWorker } from './scheduler/SchedulerWorker';
+import * as services from './services/services';
+import { wrapOtelSpan } from './utils';
+import { VERSION } from './version';
+
+// We need to override this interface to have our user typing
+declare global {
+    namespace Express {
+        interface User extends SessionUser {}
+    }
+}
+
+type AppArguments = {
+    lightdashConfig: LightdashConfig;
+    port: string | number;
+    otelSdk: NodeSDK;
+    environment?: 'production' | 'development';
+};
+
+export default class App {
+    private readonly lightdashConfig: LightdashConfig;
+
+    private readonly analytics: LightdashAnalytics;
+
+    private readonly otelSdk: NodeSDK;
+
+    private readonly port: string | number;
+
+    private readonly environment: 'production' | 'development';
+
+    private schedulerWorker: SchedulerWorker | undefined;
+
+    constructor(args: AppArguments) {
+        this.lightdashConfig = args.lightdashConfig;
+        this.otelSdk = args.otelSdk;
+        this.port = args.port;
+        this.environment = args.environment || 'production';
+        this.analytics = new LightdashAnalytics({
+            lightdashConfig: this.lightdashConfig,
+            writeKey: this.lightdashConfig.rudder.writeKey || 'notrack',
+            dataPlaneUrl: this.lightdashConfig.rudder.dataPlaneUrl
+                ? `${this.lightdashConfig.rudder.dataPlaneUrl}/v1/batch`
+                : 'notrack',
+            options: {
+                enable:
+                    this.lightdashConfig.rudder.writeKey &&
+                    this.lightdashConfig.rudder.dataPlaneUrl,
+            },
+        });
+    }
+
+    async start() {
+        // @ts-ignore
+        // eslint-disable-next-line no-extend-native, func-names
+        BigInt.prototype.toJSON = function () {
+            return this.toString();
+        };
+
+        if (this.environment !== 'development') {
+            App.initNodeProcessMonitor();
+        }
+
+        const expressApp = this.initExpress();
+        this.initSentry(expressApp);
+        this.initSlack();
+        if (this.lightdashConfig.scheduler?.enabled) {
+            this.initSchedulerWorker();
+        }
+    }
+
+    private initExpress() {
+        const expressApp = express();
+
+        const KnexSessionStore = connectSessionKnex(expressSession);
+
+        const store = new KnexSessionStore({
+            knex: database as any,
+            createtable: false,
+            tablename: 'sessions',
+            sidfieldname: 'sid',
+        });
+
+        expressApp.use(
+            express.json({ limit: this.lightdashConfig.maxPayloadSize }),
+        );
+
+        expressApp.use(bodyParser.json());
+        expressApp.use(bodyParser.urlencoded({ extended: false }));
+        expressApp.use(cookieParser());
+
+        expressApp.use(
+            expressSession({
+                secret: this.lightdashConfig.lightdashSecret,
+                proxy: this.lightdashConfig.trustProxy,
+                rolling: true,
+                cookie: {
+                    maxAge:
+                        (this.lightdashConfig.cookiesMaxAgeHours || 24) *
+                        60 *
+                        60 *
+                        1000, // in ms
+                    secure: this.lightdashConfig.secureCookies,
+                    httpOnly: true,
+                    sameSite: 'lax',
+                },
+                resave: false,
+                saveUninitialized: false,
+                store,
+            }),
+        );
+        expressApp.use(flash());
+        expressApp.use(passport.initialize());
+        expressApp.use(passport.session());
+
+        expressApp.use(expressWinstonMiddleware);
+
+        expressApp.get('/', (req, res) => {
+            res.sendFile(
+                path.join(__dirname, '../../frontend/build', 'index.html'),
+                {
+                    headers: { 'Cache-Control': 'no-cache, private' },
+                },
+            );
+        });
+        // api router
+        expressApp.use('/api/v1', apiV1Router);
+        RegisterRoutes(expressApp);
+        // Api docs
+        if (
+            this.lightdashConfig.mode === LightdashMode.PR ||
+            this.environment !== 'production'
+        ) {
+            expressApp.get('/api/docs/openapi.json', (req, res) => {
+                res.send(apiSpec);
+            });
+            expressApp.get(
+                '/api/docs',
+                reDoc({
+                    title: 'Lightdash API Docs',
+                    specUrl: '/api/docs/openapi.json',
+                }),
+            );
+        }
+
+        // frontend assets - immutable because vite appends hash to filenames
+        expressApp.use(
+            '/assets',
+            expressStaticGzip(
+                path.join(__dirname, '../../frontend/build/assets'),
+                {
+                    index: false,
+                    customCompressions: [
+                        {
+                            encodingName: 'gzip',
+                            fileExtension: 'gzip',
+                        },
+                    ],
+                },
+            ),
+        );
+        // frontend static files - no cache
+        expressApp.use(
+            express.static(path.join(__dirname, '../../frontend/build'), {
+                setHeaders: () => ({
+                    // private - browsers can cache but not CDNs
+                    // no-cache - caches must revalidate with the origin server before using a cached copy
+                    'Cache-Control': 'no-cache, private',
+                }),
+            }),
+        );
+
+        expressApp.get('*', (req, res) => {
+            res.sendFile(
+                path.join(__dirname, '../../frontend/build', 'index.html'),
+                {
+                    headers: { 'Cache-Control': 'no-cache, private' },
+                },
+            );
+        });
+
+        // Start the server
+        expressApp.listen(this.port, () => {
+            if (this.environment === 'production') {
+                Logger.info(
+                    `\n   |     |     |     |     |     |     |\n   |     |     |     |     |     |     |\n   |     |     |     |     |     |     |  \n \\ | / \\ | / \\ | / \\ | / \\ | / \\ | / \\ | /\n  \\|/   \\|/   \\|/   \\|/   \\|/   \\|/   \\|/\n------------------------------------------\nLaunch lightdash at http://localhost:${this.port}\n------------------------------------------\n  /|\\   /|\\   /|\\   /|\\   /|\\   /|\\   /|\\\n / | \\ / | \\ / | \\ / | \\ / | \\ / | \\ / | \\\n   |     |     |     |     |     |     |\n   |     |     |     |     |     |     |\n   |     |     |     |     |     |     |`,
+                );
+            }
+        });
+
+        // Errors
+        expressApp.use(Sentry.Handlers.errorHandler()); // The Sentry error handler must be before any other error middleware and after all controllers
+        expressApp.use(
+            (error: Error, req: Request, res: Response, _: NextFunction) => {
+                const errorResponse = errorHandler(error);
+                Logger.error(
+                    `Handled error of type ${errorResponse.name} on [${req.method}] ${req.path}`,
+                    errorResponse,
+                );
+                this.analytics.track({
+                    event: 'api.error',
+                    userId: req.user?.userUuid,
+                    anonymousId: !req.user?.userUuid
+                        ? LightdashAnalytics.anonymousId
+                        : undefined,
+                    properties: {
+                        name: errorResponse.name,
+                        statusCode: errorResponse.statusCode,
+                        route: req.path,
+                        method: req.method,
+                    },
+                });
+                res.status(errorResponse.statusCode).send({
+                    status: 'error',
+                    error: {
+                        statusCode: errorResponse.statusCode,
+                        name: errorResponse.name,
+                        message: errorResponse.message,
+                        data: errorResponse.data,
+                    },
+                });
+            },
+        );
+
+        // Authentication
+        passport.use(apiKeyPassportStrategy);
+        passport.use(localPassportStrategy);
+        if (googlePassportStrategy) {
+            passport.use(googlePassportStrategy);
+            refresh.use(googlePassportStrategy);
+        }
+        if (isOktaPassportStrategyAvailableToUse) {
+            passport.use('okta', new OpenIDClientOktaStrategy());
+        }
+        if (oneLoginPassportStrategy) {
+            passport.use('oneLogin', oneLoginPassportStrategy);
+        }
+        if (azureAdPassportStrategy) {
+            passport.use('azuread', azureAdPassportStrategy);
+        }
+        passport.serializeUser((user, done) => {
+            // On login (user changes), user.userUuid is written to the session store in the `sess.passport.data` field
+            done(null, user.userUuid);
+        });
+
+        // Before each request handler we read `sess.passport.user` from the session store
+        passport.deserializeUser(async (id: string, done) => {
+            // Convert to a full user profile
+            try {
+                const user = await wrapOtelSpan(
+                    'Passport.deserializeUser',
+                    {},
+                    () => userModel.findSessionUserByUUID(id),
+                );
+                // Store that user on the request (`req`) object
+                done(null, user);
+            } catch (e) {
+                done(e);
+            }
+        });
+
+        return expressApp;
+    }
+
+    private initSlack() {
+        const slackService = new SlackService({
+            slackAuthenticationModel,
+            lightdashConfig: this.lightdashConfig,
+            analytics: this.analytics,
+        });
+    }
+
+    private initSentry(expressApp: Express) {
+        Sentry.init({
+            release: VERSION,
+            dsn: this.lightdashConfig.sentry.dsn,
+            environment:
+                this.environment === 'development'
+                    ? 'development'
+                    : this.lightdashConfig.mode,
+            integrations: [
+                new Sentry.Integrations.Http({ tracing: true }),
+                new Tracing.Integrations.Express({
+                    app: expressApp,
+                }),
+            ],
+            ignoreErrors: ['WarehouseQueryError', 'FieldReferenceError'],
+            tracesSampler: (context: SamplingContext): boolean | number => {
+                if (
+                    context.request?.url?.endsWith('/status') ||
+                    context.request?.url?.endsWith('/health') ||
+                    context.request?.url?.endsWith('/favicon.ico') ||
+                    context.request?.url?.endsWith('/robots.txt') ||
+                    context.request?.url?.endsWith('livez') ||
+                    context.request?.headers?.['user-agent']?.includes(
+                        'GoogleHC',
+                    )
+                ) {
+                    return 0.0;
+                }
+                return 0.2;
+            },
+            beforeBreadcrumb(breadcrumb) {
+                if (
+                    breadcrumb.category === 'http' &&
+                    breadcrumb?.data?.url &&
+                    breadcrumb.data.url.startsWith('https://hub.docker.com')
+                ) {
+                    return null;
+                }
+                return breadcrumb;
+            },
+        });
+        expressApp.use(
+            Sentry.Handlers.requestHandler({
+                user: [
+                    'userUuid',
+                    'organizationUuid',
+                    'organizationName',
+                    'email',
+                ],
+            }) as RequestHandler,
+        );
+        expressApp.use(Sentry.Handlers.tracingHandler());
+    }
+
+    private initSchedulerWorker() {
+        this.schedulerWorker = new SchedulerWorker({
+            lightdashConfig: this.lightdashConfig,
+            analytics: this.analytics,
+            ...services,
+            ...clients,
+        });
+        this.schedulerWorker.run().catch((e) => {
+            Logger.error('Error starting scheduler worker', e);
+        });
+    }
+
+    static initNodeProcessMonitor() {
+        // Monitor Node.js process with opentelemetry
+        registerNodeMetrics();
+    }
+
+    async stop() {
+        if (this.schedulerWorker && this.schedulerWorker.runner) {
+            try {
+                await this.schedulerWorker.runner.stop();
+                Logger.info('Stopped scheduler worker');
+            } catch (e) {
+                Logger.error('Error stopping scheduler worker', e);
+            }
+        }
+        if (postHogClient) {
+            try {
+                await postHogClient.shutdownAsync();
+                Logger.info('Stopped PostHog Client');
+            } catch (e) {
+                Logger.error('Error stopping PostHog Client', e);
+            }
+        }
+        if (this.otelSdk) {
+            try {
+                await this.otelSdk.shutdown();
+                Logger.info('Stopped OpenTelemetry SDK');
+            } catch (e) {
+                Logger.error('Error stopping OpenTelemetry SDK', e);
+            }
+        }
+    }
+}
